refactors completions out of lsp plugin module and adds go lsp
