--- nvim/lua/plugins/lsp/completions/init.lua
@@ -1,95 +1,0 @@
-local M = {}
-
-M.setup = function()
-  local cmp = require("cmp")
-  local lspkind = require("lspkind")
-  local luasnip = require("luasnip")
-
-  -- Used by formatter
-  local source_mapping = {
-    buffer = "[Buffer]",
-    nvim_lsp = "[LSP]",
-    nvim_lua = "[Lua]",
-    cmp_tabnine = "[TN]",
-    cmdline = "[Cmd]",
-    path = "[Path]",
-  }
-
-  local function formatTabnine(entry, vim_item)
-    vim_item.kind = lspkind.presets.default[vim_item.kind]
-    local menu = source_mapping[entry.source.name]
-    if entry.source.name == "cmp_tabnine" then
-      local data = entry.completion_item.data
-      if data ~= nil and data.detail ~= nil then
-        menu = data.detail .. " " .. menu
-      end
-      vim_item.kind = ""
-    end
-    vim_item.menu = menu
-    return vim_item
-  end
-
-  cmp.setup({
-    snippet = {
-      expand = function(args)
-        luasnip.lsp_expand(args.body)
-      end,
-    },
-    window = {
-      -- completion = cmp.config.window.bordered(),
-      -- documentation = cmp.config.window.bordered(),
-    },
-    mapping = cmp.mapping.preset.insert({
-      ["<C-b>"] = cmp.mapping.scroll_docs(-4),
-      ["<C-f>"] = cmp.mapping.scroll_docs(4),
-      ["<C-Space>"] = cmp.mapping.complete(),
-      ["<C-e>"] = cmp.mapping.abort(),
-      ["<CR>"] = cmp.mapping.confirm(),
-      ["<Tab>"] = cmp.mapping(cmp.mapping.select_next_item(), { "i", "s" }),
-      ["<S-Tab>"] = cmp.mapping(cmp.mapping.select_prev_item(), { "i", "s" }),
-    }),
-    sources = cmp.config.sources({
-      { name = "nvim_lsp" },
-      { name = "luasnip" },
-      { name = "cmp_tabnine" },
-    }, {
-      { name = "buffer" },
-    }),
-    formatting = {
-      format = lspkind.cmp_format({
-        mode = "symbol_text",
-        maxwidth = 200,
-        before = formatTabnine,
-      }),
-    },
-  })
-
-  cmp.setup.cmdline("/", {
-    mapping = cmp.mapping.preset.cmdline(),
-    sources = {
-      { name = "buffer" },
-    },
-  })
-
-  cmp.setup.cmdline(":", {
-    mapping = cmp.mapping.preset.cmdline(),
-    sources = cmp.config.sources({
-      { name = "path" },
-    }, {
-      { name = "cmdline", keyword_length = 2 },
-    }),
-  })
-
-  -- Tabnine AI completions
-  local tabnine = require("cmp_tabnine.config")
-  tabnine:setup({
-    max_lines = 1000,
-    max_num_results = 20,
-    sort = true,
-    run_on_every_keystroke = true,
-    show_prediction_strength = true,
-    snippet_placeholder = "",
-  })
-end
-
-return M
